misread comment, fixed so string functions exist in stringUtils file
