+++ www/src/templates/Markdown/Markdown.tsx
@@ -17,8 +17,10 @@ import {Header} from '../../components/Header';
 import {Footer} from '../../components/Footer';
 import {Navigation} from '../../components/Navigation';
 import {Code} from '../../components/Code';
-import findIndexOf from '../../utils/findIndexOf';
-import replaceStringWith from '../../utils/replaceStringWith';
+import {
+	findIndexOf,
+	replaceStringWith
+} from '../../utils/stringUtils';
 
 //import './prism.css';
 import './Markdown.css';
