--- www/src/utils/replaceStringWith.ts
@@ -1,5 +1,0 @@
-// function which will replace string with arguements given.
-
-export default function replaceStringWith(str: string, replaceString: string, replaceWith: string): string {
-  return str.replace(replaceString, replaceWith);
-};
