+++ www/src/utils/tests/stringUtils.test.ts
@@ -1,4 +1,33 @@
-import findIndexOf from '../findIndexOf';
+import {
+  findIndexOf,
+  replaceStringWith
+} from '../stringUtils';
+
+describe('replaceStringWith.spec', () => {
+  
+  const testString = 'www.thisisatest.com/testing/some/stuff';
+
+  it('should replace the string with the arguement given', () => {
+
+    const expected = 'www.thisisatest.com';
+
+    const actual: string = replaceStringWith(testString, '/testing/some/stuff', '');
+
+    expect(actual).toBe(expected);
+
+  });
+
+  it('should return same string if search term not found', () => {
+
+    const expected: string = testString;
+
+    const actual: string = replaceStringWith(testString, 'booop', '');
+
+    expect(actual).toBe(expected);
+
+  });
+
+});
 
 describe('findIndexOf.spec', () => {
